add map_lib.rs
