+++ source/pervasive/map.rs
@@ -6,8 +6,6 @@ use builtin_macros::*;
 use crate::pervasive::*;
 use crate::set::*;
 use core::marker;
-use crate::set_lib::lemma_set_properties;
-
 
 verus! {
 
@@ -117,317 +115,6 @@ impl<K, V> Map<K, V> {
     pub open spec fn ext_equal(self, m2: Map<K, V>) -> bool {
         self =~= m2
     }
-
-    /// Returns true if the key `k` is in the domain of `self`.
-
-    #[verifier(inline)]
-    pub open spec fn contains_key(self, k: K) -> bool {
-        self.dom().contains(k)
-    }
-
-    /// Returns true if the value `v` is in the map of `self`.
-
-    pub open spec fn contains_value(self, v: V) -> bool {
-        exists|i: K| #[trigger] self.dom().contains(i) && self[i] == v
-    }
-
-    /// Returns true if the key `k` is in the domain of `self`, and it maps to the value `v`.
-
-    pub open spec fn contains_pair(self, k: K, v: V) -> bool {
-        self.dom().contains(k) && self[k] == v
-    }
-
-    /// Returns true if `m1` is _contained in_ `m2`, i.e., the domain of `m1` is a subset
-    /// of the domain of `m2`, and they agree on all values in `m1`.
-    ///
-    /// ## Example
-    ///
-    /// ```rust
-    /// assert(
-    ///    map![1 => 10, 2 => 11].le(map![1 => 10, 2 => 11, 3 => 12])
-    /// );
-    /// ```
-
-    pub open spec fn le(self, m2: Self) -> bool {
-        forall|k: K| #[trigger] self.dom().contains(k) ==>
-            #[trigger] m2.dom().contains(k) && self[k] == m2[k]
-    }
-
-    /// Gives the union of two maps, defined as:
-    ///  * The domain is the union of the two input maps.
-    ///  * For a given key in _both_ input maps, it maps to the same value that it maps to in the _right_ map (`m2`).
-    ///  * For any other key in either input map (but not both), it maps to the same value
-    ///    as it does in that map.
-    ///
-    /// ## Example
-    ///
-    /// ```rust
-    /// assert_maps_equal!(
-    ///    map![1 => 10, 2 => 11].union_prefer_right(map![1 => 20, 3 => 13]),
-    ///    map![1 => 20, 2 => 11, 3 => 13],
-    /// );
-    /// ```
-
-    pub open spec fn union_prefer_right(self, m2: Self) -> Self {
-        Self::new(
-            |k: K| self.dom().contains(k) || m2.dom().contains(k),
-            |k: K| if m2.dom().contains(k) { m2[k] } else { self[k] }
-        )
-    }
-
-    /// Removes the given keys and their associated values from the map.
-    ///
-    /// Ignores any key in `keys` which is not in the domain of `self`.
-    ///
-    /// ## Example
-    ///
-    /// ```rust
-    /// assert_maps_equal!(
-    ///    map![1 => 10, 2 => 11, 3 => 12].remove_keys(set!{2, 3, 4}),
-    ///    map![1 => 10],
-    /// );
-    /// ```
-
-    pub open spec fn remove_keys(self, keys: Set<K>) -> Self {
-        Self::new(
-            |k: K| self.dom().contains(k) && !keys.contains(k),
-            |k: K| self[k]
-        )
-    }
-
-    /// Complement to `remove_keys`. Restricts the map to (key, value) pairs
-    /// for keys that are _in_ the given set; that is, it removes any keys
-    /// _not_ in the set.
-    ///
-    /// ## Example
-    ///
-    /// ```rust
-    /// assert_maps_equal!(
-    ///    map![1 => 10, 2 => 11, 3 => 12].remove_keys(set!{2, 3, 4}),
-    ///    map![2 => 11, 3 => 12],
-    /// );
-    /// ```
-
-    pub open spec fn restrict(self, keys: Set<K>) -> Self {
-        Self::new(
-            |k: K| self.dom().contains(k) && keys.contains(k),
-            |k: K| self[k]
-        )
-    }
-
-    /// Returns `true` if and only if the given key maps to the same value or does not exist in self and m2.
-    
-    pub open spec fn equal_on_key(self, m2: Self, key: K) -> bool
-    {
-        ||| (!self.dom().contains(key) && !m2.dom().contains(key))
-        ||| (self.dom().contains(key) && m2.dom().contains(key) && self[key] == m2[key])
-    }
-
-    /// Returns `true` if the two given maps agree on all keys that their domains
-    /// share in common.
-
-    pub open spec fn agrees(self, m2: Self) -> bool {
-        forall|k| #![auto] self.dom().contains(k) && m2.dom().contains(k) ==>
-            self[k] == m2[k]
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_empty() -> (tracked out_v: Self)
-        ensures
-            out_v == Map::<K, V>::empty(),
-    {
-        unimplemented!();
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_insert(tracked &mut self, key: K, tracked value: V)
-        ensures
-            *self == Map::insert(*old(self), key, value),
-    {
-        unimplemented!();
-    }
-
-    /// todo fill in documentation
-
-    #[verifier(external_body)]
-    pub proof fn tracked_remove(tracked &mut self, key: K) -> (tracked v: V)
-        requires
-            old(self).dom().contains(key),
-        ensures
-            *self == Map::remove(*old(self), key),
-            v == old(self)[key],
-    {
-        unimplemented!();
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_borrow(tracked &self, key: K) -> (tracked v: &V)
-        requires
-            self.dom().contains(key),
-        ensures
-            *v === self.index(key),
-    {
-        unimplemented!();
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_map_keys<J>(
-        tracked old_map: Map<K, V>,
-        key_map: Map<J, K>
-    ) -> (tracked new_map: Map<J, V>)
-        requires
-            forall |j| #![auto] key_map.dom().contains(j) ==> old_map.dom().contains(key_map.index(j)),
-            forall |j1, j2| #![auto]
-                !equal(j1, j2) && key_map.dom().contains(j1) && key_map.dom().contains(j2)
-                ==> !equal(key_map.index(j1), key_map.index(j2))
-        ensures
-            forall |j| #[trigger] new_map.dom().contains(j) <==> key_map.dom().contains(j),
-            forall |j| key_map.dom().contains(j) ==>
-                new_map.dom().contains(j) &&
-                #[trigger] new_map.index(j) == old_map.index(key_map.index(j)),
-    {
-        unimplemented!();
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_remove_keys(tracked &mut self, keys: Set<K>)
-        -> (tracked out_map: Map<K, V>)
-        requires
-            keys.subset_of(old(self).dom())
-        ensures
-            self == old(self).remove_keys(keys),
-            out_map == old(self).restrict(keys),
-    {
-        unimplemented!();
-    }
-
-    #[verifier(external_body)]
-    pub proof fn tracked_union_prefer_right(tracked &mut self, right: Self)
-        ensures
-            *self == old(self).union_prefer_right(right),
-    {
-        unimplemented!();
-    }
-
-    /// Map a function `f` over all (k, v) pairs in `self`.
-    pub open spec fn map_entries<W>(self, f: FnSpec(K, V) -> W) -> Map<K, W> {
-        Map::new(|k: K| self.contains_key(k), |k: K| f(k, self[k]))
-    }
-
-    /// Map a function `f` over the values in `self`.
-    pub open spec fn map_values<W>(self, f: FnSpec(V) -> W) -> Map<K, W> {
-        Map::new(|k: K| self.contains_key(k), |k: K| f(self[k]))
-    }
-
-    /// Returns `true` if and only if a map is injective
-    pub open spec fn is_injective(self) -> bool {
-        forall |x: K, y: K| x != y && self.dom().contains(x) && self.dom().contains(y) ==> #[trigger] self[x] != #[trigger] self[y]
-    }
-
-    /// Swaps map keys and values. Values are not required to be unique; no
-    /// promises on which key is chosen on the intersection.
-    pub open spec fn invert(self) -> Map<V,K> 
-    {
-        Map::<V,K>::new(
-            |v: V| self.contains_value(v),
-            |v: V| choose |k: K| self.contains_pair(k,v)
-        )
-    }
-}
-
-impl Map<int,int> {
-
-    /// Returns `true` if a map is monotonic -- that is, if the mapping between ordered sets 
-    /// preserves the regular `<=` ordering on integers.
-    pub open spec fn is_monotonic(self) -> bool {
-        forall |x: int, y: int| self.dom().contains(x) && self.dom().contains(y) && x <= y 
-            ==> #[trigger] self[x] <= #[trigger] self[y]
-    }
-
-    /// Returns `true` if and only if a map is monotonic, only considering keys greater than
-    /// or equal to start
-    pub open spec fn monotonic_from(self, start: int) -> bool {
-        forall |x: int, y: int| self.dom().contains(x) && self.dom().contains(y) && start <= x <= y 
-            ==> #[trigger] self[x] <= #[trigger] self[y]
-    }
-
-}
-
-// Proven lemmas
-
-/// Removing a key from a map that previously contained that key results
-/// in a length that is one less than the original length.
-pub proof fn lemma_remove_key_len<K,V>(m: Map<K,V>, key: K)
-    requires
-        m.dom().contains(key),
-        m.dom().finite(),
-    ensures
-        m.dom().len() == 1 + m.remove(key).dom().len(),
-{}
-
-/// The domain of a map after removing a key is equivalent to removing the key from 
-/// the domain of the original map.
-pub proof fn lemma_remove_equivalency<K,V>(m: Map<K,V>, key: K)
-    ensures
-        m.remove(key).dom() == m.dom().remove(key),
-{}
-
-/// Removing a set of n keys from a map that previously contained all n keys
-/// results in a domain of size n less than the original domain.
-pub proof fn lemma_remove_keys_len<K,V>(m: Map<K,V>, keys: Set<K>)
-    requires
-        forall |k: K| #[trigger] keys.contains(k) ==> m.contains_key(k),
-        keys.finite(),
-        m.dom().finite(),
-    ensures 
-        m.remove_keys(keys).dom().len() == m.dom().len() - keys.len(),
-    decreases
-        keys.len(),
-{
-    lemma_set_properties::<K>();
-    if keys.len() > 0 {
-        let key = keys.choose();
-        let val = m[key];
-        lemma_remove_keys_len(m.remove(key),keys.remove(key));
-        assert(m.remove(key).remove_keys(keys.remove(key)) =~= m.remove_keys(keys));
-    }
-    else {
-        assert(m.remove_keys(keys) =~= m);
-    }
-}
-
-/// The size of the union of two maps is equal to the sum of the sizes of the individual maps
-pub proof fn lemma_disjoint_union_size<K,V>(m1: Map<K,V>, m2: Map<K,V>)
-    requires
-        m1.dom().disjoint(m2.dom()),
-        m1.dom().finite(),
-        m2.dom().finite(),
-    ensures
-        m1.union_prefer_right(m2).dom().len() == m1.dom().len() + m2.dom().len(),
-{
-    let u = m1.union_prefer_right(m2);
-    assert(u.dom() =~= m1.dom() + m2.dom()); //proves u.dom() is finite
-    assert(u.remove_keys(m1.dom()).dom() =~= m2.dom());
-    assert(u.remove_keys(m1.dom()).dom().len() == u.dom().len() - m1.dom().len()) by {
-        lemma_remove_keys_len(u, m1.dom());
-    }
-}
-
-/// The function `invert` results in an injective map
-pub proof fn lemma_invert_is_injective<K,V>(m: Map<K,V>)
-    ensures
-        m.invert().is_injective(),
-{
-    assert forall |x: V, y: V| x != y && m.invert().dom().contains(x) && m.invert().dom().contains(y) 
-                implies #[trigger] m.invert()[x] != #[trigger] m.invert()[y] by {
-        let i = choose |i: K| #[trigger] m.dom().contains(i) && m[i] == x;
-        assert(m.contains_pair(i,x));
-        let j = choose |j: K| m.contains_pair(j,x) && m.invert()[x] == j;
-
-        let k = choose |k: K| #[trigger] m.dom().contains(k) && m[k] == y;
-        assert(m.contains_pair(k,y));
-        let l = choose |l: K| m.contains_pair(l,y) && m.invert()[y] == l && l != j;
-    }   
 }
 
 // Trusted axioms
@@ -530,32 +217,6 @@ pub proof fn axiom_map_remove_different<K, V>(m: Map<K, V>, key1: K, key2: K)
 {
 }
 
-// Ported from Dafny prelude
-/// The domain of a map constructed with `Map::new(fk, fv)` is equivalent to the set constructed with `Set::new(fk)`.
-pub proof fn lemma_map_new_domain<K,V>(fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V)
-    ensures
-        #[trigger] Map::<K,V>::new(fk,fv).dom() == Set::<K>::new(|k: K| fk(k))
-{
-    assert(Set::new(fk) =~= Set::<K>::new(|k: K| fk(k)));
-}
-
-// Ported from Dafny prelude
-/// The set of values of a map constructed with `Map::new(fk, fv)` is equivalent to 
-/// the set constructed with `Set::new(|v: V| (exists |k: K| fk(k) && fv(k) == v)`. In other words,
-/// the set of all values fv(k) where fk(k) is true.
-pub proof fn lemma_map_new_values<K,V>(fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V)
-    ensures
-        #[trigger] Map::<K,V>::new(fk,fv).values() == Set::<V>::new(|v: V| (exists |k: K| #[trigger] fk(k) && #[trigger] fv(k) == v)),
-{
-    let keys = Set::<K>::new(fk);
-    let values = Map::<K,V>::new(fk,fv).values();
-    let map = Map::<K,V>::new(fk,fv);
-
-    assert(map.dom() =~= keys);
-    assert(forall |k: K| #[trigger] fk(k) ==> keys.contains(k));
-    assert(values =~= Set::<V>::new(|v: V| (exists |k: K| #[trigger] fk(k) && #[trigger] fv(k) == v)));
-}
-
 /// Two maps are equivalent if their domains are equivalent and every key in their domains map to the same value.
 #[verifier(external_body)]
 #[verifier(broadcast_forall)]
@@ -577,24 +238,6 @@ pub proof fn axiom_map_ext_equal_deep<K, V>(m1: Map<K, V>, m2: Map<K, V>)
             &&& forall|k: K| #![auto] m1.dom().contains(k) ==> m1[k] =~~= m2[k]
         },
 {
-}
-
-// magic auto style bundle of lemmas that Dafny considers when proving properties of maps
-pub proof fn lemma_map_properties<K,V>()
-    ensures
-    forall |fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V| #[trigger] Map::<K,V>::new(fk,fv).dom()
-            == Set::<K>::new(|k: K| fk(k)), //lemma_map_new_domain
-    forall |fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V| #[trigger] Map::<K,V>::new(fk,fv).values() 
-            == Set::<V>::new(|v: V| exists |k: K| #[trigger] fk(k) && #[trigger] fv(k) == v),  //lemma_map_new_values
-{
-    assert forall |fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V| 
-        #[trigger] Map::<K,V>::new(fk,fv).dom() == Set::<K>::new(|k: K| fk(k)) by {
-            lemma_map_new_domain(fk, fv);
-        }
-    assert forall |fk: FnSpec(K) -> bool, fv: FnSpec(K) -> V| #[trigger] Map::<K,V>::new(fk,fv).values() 
-        == Set::<V>::new(|v: V| exists |k: K| #[trigger] fk(k) && #[trigger] fv(k) == v) by {
-            lemma_map_new_values(fk, fv);
-        }
 }
 
 // Macros
